moved type converter implementations into own module
