+++ livingdoc-fixture-api/build.gradle
@@ -1,28 +1,6 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'org.junit.platform:junit-platform-gradle-plugin:1.0.0-M4'
-    }
-}
-
 apply plugin: 'java'
-apply plugin: 'org.junit.platform.gradle.plugin'
 
 sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testCompile 'org.assertj:assertj-core:3.6.2'
-    testCompile 'org.mockito:mockito-core:2.7.22'
-    testCompile 'org.junit.jupiter:junit-jupiter-api:5.0.0-M4'
-    testCompile 'org.junit.jupiter:junit-jupiter-params:5.0.0-M4'
-    testRuntime 'org.junit.jupiter:junit-jupiter-engine:5.0.0-M4'
-}
 
 task javadocJar(type: Jar) {
     classifier = 'javadoc'
