+++ livingdoc-fixture-api/src/main/java/org/livingdoc/fixture/api/converter/ConversionException.java
@@ -1,0 +1,17 @@
+package org.livingdoc.fixture.api.converter;
+
+public class ConversionException extends RuntimeException {
+
+    public ConversionException(String message) {
+        super(message);
+    }
+
+    public ConversionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ConversionException(Throwable cause) {
+        super(cause);
+    }
+
+}
