--- livingdoc-fixture-api/src/main/java/org/livingdoc/fixture/api/converter/common/LongConverter.java
@@ -1,21 +1,0 @@
-package org.livingdoc.fixture.api.converter.common;
-
-import java.lang.reflect.AnnotatedElement;
-
-import org.livingdoc.fixture.api.converter.exceptions.ConversionException;
-import org.livingdoc.fixture.api.converter.TypeConverter;
-
-
-public class LongConverter implements TypeConverter<Long> {
-
-    @Override
-    public Long convert(String value, AnnotatedElement element) throws ConversionException {
-        try {
-            String trimmedValue = value.trim();
-            return Long.valueOf(trimmedValue);
-        } catch (NumberFormatException e) {
-            throw new ConversionException(e);
-        }
-    }
-
-}
