--- livingdoc-fixture-api/src/main/java/org/livingdoc/fixture/api/converter/exceptions/ConversionException.java
@@ -1,17 +1,0 @@
-package org.livingdoc.fixture.api.converter.exceptions;
-
-public class ConversionException extends RuntimeException {
-
-    public ConversionException(String message) {
-        super(message);
-    }
-
-    public ConversionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ConversionException(Throwable cause) {
-        super(cause);
-    }
-
-}
