--- livingdoc-fixture-api/src/main/java/org/livingdoc/fixture/api/converter/exceptions/MalformedFormatException.java
@@ -1,11 +1,0 @@
-package org.livingdoc.fixture.api.converter.exceptions;
-
-public class MalformedFormatException extends RuntimeException {
-
-    private static final String MESSAGE = "Custom format pattern is malformed: %s";
-
-    public MalformedFormatException(Throwable cause) {
-        super(String.format(MESSAGE, cause.getMessage()), cause);
-    }
-
-}
