+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/common/StringConverter.java
@@ -1,0 +1,15 @@
+package org.livingdoc.fixture.converter.common;
+
+import java.lang.reflect.AnnotatedElement;
+
+import org.livingdoc.fixture.api.converter.TypeConverter;
+
+
+public class StringConverter implements TypeConverter<String> {
+
+    @Override
+    public String convert(String value, AnnotatedElement element) {
+        return value;
+    }
+
+}
