+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/exceptions/MalformedFormatException.java
@@ -1,0 +1,11 @@
+package org.livingdoc.fixture.converter.exceptions;
+
+public class MalformedFormatException extends RuntimeException {
+
+    private static final String MESSAGE = "Custom format pattern is malformed: %s";
+
+    public MalformedFormatException(Throwable cause) {
+        super(String.format(MESSAGE, cause.getMessage()), cause);
+    }
+
+}
