+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/exceptions/ValueFormatException.java
@@ -1,0 +1,14 @@
+package org.livingdoc.fixture.converter.exceptions;
+
+import org.livingdoc.fixture.api.converter.ConversionException;
+
+
+public class ValueFormatException extends ConversionException {
+
+    private static final String MESSAGE = "The value '%s' does not match the expected format: %s";
+
+    public ValueFormatException(String value, String format, Throwable cause) {
+        super(String.format(MESSAGE, value, format), cause);
+    }
+
+}
