+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/time/LocalDateConverter.java
@@ -1,0 +1,19 @@
+package org.livingdoc.fixture.converter.time;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+
+public class LocalDateConverter extends AbstractTemporalConverter<LocalDate> {
+
+    @Override
+    DateTimeFormatter getDefaultFormatter() {
+        return DateTimeFormatter.ISO_DATE;
+    }
+
+    @Override
+    LocalDate parse(String value, DateTimeFormatter formatter) {
+        return LocalDate.parse(value, formatter);
+    }
+
+}
