+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/time/LocalDateTimeConverter.java
@@ -1,0 +1,19 @@
+package org.livingdoc.fixture.converter.time;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+
+public class LocalDateTimeConverter extends AbstractTemporalConverter<LocalDateTime> {
+
+    @Override
+    DateTimeFormatter getDefaultFormatter() {
+        return DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+    }
+
+    @Override
+    LocalDateTime parse(String value, DateTimeFormatter formatter) {
+        return LocalDateTime.parse(value, formatter);
+    }
+
+}
