+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/time/LocalTimeConverter.java
@@ -1,0 +1,19 @@
+package org.livingdoc.fixture.converter.time;
+
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+
+
+public class LocalTimeConverter extends AbstractTemporalConverter<LocalTime> {
+
+    @Override
+    DateTimeFormatter getDefaultFormatter() {
+        return DateTimeFormatter.ISO_TIME;
+    }
+
+    @Override
+    LocalTime parse(String value, DateTimeFormatter formatter) {
+        return LocalTime.parse(value, formatter);
+    }
+
+}
